New system test for websocket
